PFCF-325: initial changes to add instructions for PO eFiled applications
